Rename <SwitchFieldGroup> to <CheckboxFieldGroup>
