--- src/frontend/web_application/src/components/form/Checkbox/style.scss
@@ -1,9 +1,0 @@
-@import '../../../styles/common';
-@import '../../../styles/vendor/bootstrap_foundation-sites';
-
-.m-checkbox {
-
-  &__label {
-    padding-left: map-get($co-form__spacing, 'small');
-  }
-}
