+++ src/frontend/web_application/src/components/form/CheckboxFieldGroup/components/Checkbox/index.jsx
