+++ src/frontend/web_application/src/components/form/CheckboxFieldGroup/components/Switch/index.jsx
