+++ src/frontend/web_application/src/components/form/CheckboxFieldGroup/components/Switch/style.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/common';
-@import '../../../styles/vendor/bootstrap_foundation-sites';
+@import '../../../../../styles/common';
+@import '../../../../../styles/vendor/bootstrap_foundation-sites';
 
 $switch-height: 2rem !default;
