+++ src/frontend/web_application/src/components/form/CheckboxFieldGroup/style.scss
