+++ src/frontend/web_application/src/components/form/index.js
@@ -3,18 +3,15 @@ import FieldErrors from './FieldErrors';
 import Fieldset, { Legend } from './Fieldset';
 import FormGrid, { FormRow, FormColumn } from './FormGrid';
 import InputText from './InputText';
-import Checkbox from './Checkbox';
 import PasswordStrength from './PasswordStrength';
 import RadioFieldGroup from './RadioFieldGroup';
 import SelectFieldGroup from './SelectFieldGroup';
-import Switch from './Switch';
-import SwitchFieldGroup from './SwitchFieldGroup';
+import CheckboxFieldGroup from './CheckboxFieldGroup';
 import TextFieldGroup from './TextFieldGroup';
 import TextareaFieldGroup from './TextareaFieldGroup';
 
 export {
   // CheckboxField,
-  Checkbox,
   FieldErrors,
   Fieldset,
   Legend,
@@ -25,8 +22,7 @@ export {
   RadioFieldGroup,
   PasswordStrength,
   SelectFieldGroup,
-  SwitchFieldGroup,
-  Switch,
+  CheckboxFieldGroup,
   TextFieldGroup,
   TextareaFieldGroup,
 };
